refactor(paging): the shared page carries a plain int, and two more caps land
The lint caught five int -> int32 narrowings in the page adapters -- offsets no
Bitbucket will ever return, each one a suppression waiting to be written. The
shared Page now carries NextPageStart as an int, so every adapter widens and
none narrows: openapi.Offset lifts the generated client's *int32, and the
endpoints decoded by hand pass their own counted offset straight through.

Two caps that were renamed last commit but not wired: reposettings' permission
listings took maxResults and then handed a constant to the walk, so the value
was ineffectual -- which is exactly what the linter said. They cap now, and the
package exports AllResults for the callers that want everything.

Those callers are the dry runs. `permission grant --dry-run` read the first 100
entries to decide create-or-update, which was harmless while the listing ignored
the number and is not harmless now: a project or repository past 100 entries
would predict "create" for a subject that already has a permission. Both grant
and revoke previews ask for all of them, the same trap AllResults was added for
in #470.

Also drops a second orphaned test helper the linter found, and a conversion in a
live test that stopped being a conversion when the comment state's type changed.

+++ internal/openapi/paging_test.go
@@ -40,9 +40,9 @@ func (source *pagedSource) fetch(_ context.Context, start, limit int) (openapi.P
 	last := end >= len(source.items)
 	page.IsLastPage = &last
 	if !last {
-		next := int32(end)
+		next := end
 		if source.stuckAt > 0 && end >= source.stuckAt {
-			next = int32(start)
+			next = start
 		}
 		page.NextPageStart = &next
 	}
@@ -217,11 +217,11 @@ func TestPageThroughLooksPastAnEmptyPage(t *testing.T) {
 		switch start {
 		case 0:
 			// Sized at 25 and then filtered down to nothing.
-			next := int32(25)
+			next := 25
 
 			return openapi.Page[int]{IsLastPage: &notLast, NextPageStart: &next}, nil
 		case 25:
-			next := int32(50)
+			next := 50
 
 			return openapi.Page[int]{Values: items(3), IsLastPage: &notLast, NextPageStart: &next}, nil
 		default:
@@ -250,7 +250,7 @@ func TestPageThroughStopsAtAnEmptyLastPage(t *testing.T) {
 		served++
 		notLast := false
 		// Says there is more, but never advances.
-		stuck := int32(0)
+		stuck := 0
 
 		return openapi.Page[int]{IsLastPage: &notLast, NextPageStart: &stuck}, nil
 	}
